restructure xgboost module
