+++ tests/xgboost/test_export_tree_data.py
@@ -1,7 +1,7 @@
 import xgboost as xgb
 
 from tabular_trees.trees import export_tree_data
-from tabular_trees.xgboost import XGBoostTabularTrees
+from tabular_trees.xgboost.xgboost_tabular_trees import XGBoostTabularTrees
 
 
 def test_model_specific_function_dispatch(xgb_diabetes_model):
